Single Qubit implementation of Deutsch algorithm
I tried to comment on everything and make it as readable as possible, if anything is confusing let me know

I implemented deutsch algorithm as a function that takes in an oracle and composes them together. The deutsch algorithm measures the first Qubit, collapsing the state.

matplotlib is required for drawing the program
pylatexenc is required for visualizing the program